Split JWT code to authn module
