+++ authn/jwt.go
@@ -1,4 +1,4 @@
-package doorman
+package authn
 
 import (
 	"fmt"
