+++ authn/jwt_claims_mozilla_test.go
@@ -1,4 +1,4 @@
-package doorman
+package authn
 
 import (
 	"testing"
