+++ authn/jwt_generic_test.go
@@ -1,4 +1,4 @@
-package doorman
+package authn
 
 import (
 	"net/http"
